langGrap-info-create/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-tavily>=0.1.0
langgraph>=0.3.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
psutil>=5.9.0 